# ui/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tawreed",
    "version": "0.1.0"
  },
  "paths": {
...
    "/projects/{project_id}/sources": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Add Sources",
        "description": "Add more BOQ files to a project. A file the project already has is ignored.",
        "operationId": "add_sources_projects__project_id__sources_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_add_sources_projects__project_id__sources_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/sources/{source_id}": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Get Source",
        "operationId": "get_source_projects__project_id__sources__source_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/sources/{source_id}/pages/{number}/layout": {
      "put": {
        "tags": [
          "sources"
        ],
        "summary": "Lay Out Sheet",
        "description": "The engineer sets a sheet's columns. Its items are read again from the cells; the AI leaves it as it is.",
        "operationId": "lay_out_sheet_projects__project_id__sources__source_id__pages__number__layout_put",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          },
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Number"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SheetLayout"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/sources/{source_id}/pages/{number}/set-aside": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Set Aside",
        "description": "The engineer says a page lists no items.",
        "operationId": "set_aside_projects__project_id__sources__source_id__pages__number__set_aside_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          },
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Number"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAside"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/sources/{source_id}/pages/{number}": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Get Page",
        "description": "A page's content. Sheets come a window of rows at a time, from row `start`.",
        "operationId": "get_page_projects__project_id__sources__source_id__pages__number__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          },
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Number"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Start"
            }
          },
          {
            "name": "count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Count"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SheetView"
                    },
                    {
                      "$ref": "#/components/schemas/TextView"
                    },
                    {
                      "$ref": "#/components/schemas/ImageView"
                    }
                  ],
                  "title": "Response Get Page Projects  Project Id  Sources  Source Id  Pages  Number  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/sources/{source_id}/pages/{number}/image": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Page Image",
        "description": "A PNG of a PDF page or an image; drawn once, then kept beside the page's content.",
        "operationId": "page_image_projects__project_id__sources__source_id__pages__number__image_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          },
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Number"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_add_sources_projects__project_id__sources_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_add_sources_projects__project_id__sources_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Handled": {
        "properties": {
          "by": {
            "type": "string",
            "enum": [
              "agent",
              "engineer"
            ],
            "title": "By"
          },
          "set_aside": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Set Aside"
          },
          "sheet": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SheetLayout"
              },
              {
                "type": "null"
              }
            ]
          },
          "items": {
            "type": "integer",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "by",
          "set_aside",
          "sheet",
          "items"
        ],
        "title": "Handled",
        "description": "How a page was read: laid out (a sheet's columns, when it is a sheet) or set aside, and by whom."
      },
      "HandledPage": {
        "properties": {
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "kind": {
            "type": "string",
            "enum": [
              "sheet",
              "page",
              "image"
            ],
            "title": "Kind"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "has_text": {
            "type": "boolean",
            "title": "Has Text"
          },
          "hidden": {
            "type": "boolean",
            "title": "Hidden"
          },
          "rows": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows"
          },
          "cols": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cols"
          },
          "width": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "handled": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Handled"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "number",
          "kind",
          "name",
          "has_text",
          "hidden",
          "rows",
          "cols",
          "width",
          "height",
          "handled"
        ],
        "title": "HandledPage"
      },
      "ImageView": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "image",
            "title": "Kind"
          },
          "width": {
            "type": "integer",
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "title": "Height"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "width",
          "height"
        ],
        "title": "ImageView"
      },
...
      "ProjectOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceOut"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at",
          "sources"
        ],
        "title": "ProjectOut"
      },
...
      "SetAside": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "SetAside"
      },
...
      "SheetLayout": {
        "properties": {
          "first_row": {
            "type": "integer",
            "minimum": 1.0,
            "title": "First Row"
          },
          "last_row": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Row"
          },
          "code": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z]{1,3}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "description": {
            "items": {
              "type": "string",
              "pattern": "^[A-Za-z]{1,3}$"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 1,
            "title": "Description"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z]{1,3}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "quantity": {
            "type": "string",
            "pattern": "^[A-Za-z]{1,3}$",
            "title": "Quantity"
          },
          "rate": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z]{1,3}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate"
          },
          "amount": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z]{1,3}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z]{1,3}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "first_row",
          "description",
          "quantity"
        ],
        "title": "SheetLayout",
        "description": "Which rows hold items and which column holds what. Columns are Excel letters; rows count from 1."
      },
      "SheetView": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "sheet",
            "title": "Kind"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "first_row": {
            "type": "integer",
            "title": "First Row"
          },
          "rows": {
            "items": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "boolean"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "type": "array",
            "title": "Rows"
          },
          "total_rows": {
            "type": "integer",
            "title": "Total Rows"
          },
          "merged": {
            "items": {
              "items": {
                "type": "integer"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Merged"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "name",
          "first_row",
          "rows",
          "total_rows",
          "merged"
        ],
        "title": "SheetView"
      },
      "SourceDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "kind": {
            "type": "string",
            "enum": [
              "spreadsheet",
              "csv",
              "pdf",
              "image"
            ],
            "title": "Kind"
          },
          "added_at": {
            "type": "string",
            "format": "date-time",
            "title": "Added At"
          },
          "status": {
            "type": "string",
            "enum": [
              "added",
              "reading",
              "read",
              "failed"
            ],
            "title": "Status"
          },
          "problem": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/HandledPage"
            },
            "type": "array",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "size",
          "kind",
          "added_at",
          "status",
          "problem",
          "page_count",
          "pages"
        ],
        "title": "SourceDetail"
      },
      "SourceOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "kind": {
            "type": "string",
            "enum": [
              "spreadsheet",
              "csv",
              "pdf",
              "image"
            ],
            "title": "Kind"
          },
          "added_at": {
            "type": "string",
            "format": "date-time",
            "title": "Added At"
          },
          "status": {
            "type": "string",
            "enum": [
              "added",
              "reading",
              "read",
              "failed"
            ],
            "title": "Status"
          },
          "problem": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "size",
          "kind",
          "added_at",
          "status",
          "problem",
          "page_count"
        ],
        "title": "SourceOut"
      },
      "TextLine": {
        "properties": {
          "top": {
            "type": "number",
            "title": "Top"
          },
          "bottom": {
            "type": "number",
            "title": "Bottom"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "top",
          "bottom",
          "text"
        ],
        "title": "TextLine"
      },
      "TextView": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "page",
            "title": "Kind"
          },
          "width": {
            "type": "number",
            "title": "Width"
          },
          "height": {
            "type": "number",
            "title": "Height"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/TextLine"
            },
            "type": "array",
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "width",
          "height",
          "lines"
        ],
        "title": "TextView"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}